# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Samadhan AI — Grievance Redressal API
  version: 0.1.0
  description: |
    REST API for the Arunachal Pradesh State Grievance Redressal Platform.
    AI-powered complaint routing, multi-department RBAC, and citizen self-service.
  contact:
    name: Samadhan AI Team
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://your-app.vercel.app/api
    description: Production (Vercel)
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '200': { description: Login successful — sets access_token + refresh_token cookies }
        '401': { description: Invalid credentials }
        '429': { description: Rate limited }
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      responses:
        '200': { description: New access_token cookie set }
        '401': { description: Invalid or expired refresh token }
  /auth/rotate-password:
    post:
      tags: [Auth]
      summary: Change admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        '200': { description: Password changed }
...
  /citizen/auth/send-otp:
    post:
      tags: [Citizen Auth]
      summary: Send OTP for login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              properties:
                identifier: { type: string, description: Phone or email }
      responses:
        '200': { description: OTP sent }
...
  /complaints/{id}/escalate:
    post:
      tags: [Complaints]
      summary: Escalate complaint
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Escalated }
...
  /citizen/complaints/track:
    get:
      tags: [Citizen]
      summary: Track complaint by ID (no auth required)
      parameters:
        - { name: id, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: Complaint status }
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT access token (admin)
    citizenCookieAuth:
      type: apiKey
      in: cookie
      name: citizen_access_token
      description: JWT access token (citizen)
    cronSecret:
      type: http
      scheme: bearer
      description: Vercel CRON_SECRET for cron endpoints